qemu: only disable shutdown and reboot when debugging

+++ build/QemuStep.zig
@@ -101,10 +101,10 @@ fn make(step: *Step, prog_node: *std.Progress.Node) !void {
     run_qemu.stdio = .inherit;
 
     // no reboot
-    run_qemu.addArg("-no-reboot");
+    if (self.options.qemu_remote_debug) run_qemu.addArg("-no-reboot");
 
     // no shutdown
-    run_qemu.addArg("-no-shutdown");
+    if (self.options.qemu_remote_debug) run_qemu.addArg("-no-shutdown");
 
     // no default devices
     run_qemu.addArg("-nodefaults");
